requirements.txt
attr==0.3.2
black==25.1.0
colorama==0.4.6
ConfigParser==7.2.0
cryptography==44.0.2
docutils==0.21.2
filelock==3.18.0
flake8==7.1.2
HTMLParser==0.0.2
ipython==8.26.0
ipywidgets==8.1.5
keyring==25.6.0
ollama==0.4.7
openai==1.74.0
Pillow==11.2.1
protobuf==6.30.2
pyOpenSSL==25.0.0
pytest==8.2.0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.0.8
Sphinx==8.2.3
thread==2.0.5
urllib3_secure_extra==0.1.0
pydantic
pydantic-settings==2.9.1
sqlmodel==0.0.24
psycopg2==2.9.10
psycopg2-binary==2.9.10
minio==7.2.15
